requirements.txt
python-pptx
Pillow
matplotlib